+++ parlai/core/params.py
@@ -73,7 +73,7 @@ class ParlaiParser(argparse.ArgumentParser):
         - add_model_args (default False) initializes the default arguments for
         loading models, including initializing arguments from that model.
         """
-        super().__init__(description='ParlAI parser.')
+        super().__init__(description='ParlAI parser.', allow_abbrev=False)
         self.register('type', 'bool', str2bool)
         self.register('type', 'class', str2class)
         self.parlai_home = (os.path.dirname(os.path.dirname(os.path.dirname(
@@ -227,6 +227,10 @@ class ParlaiParser(argparse.ArgumentParser):
             help='number of threads. If batchsize set to 1, used for hogwild; '
                  'otherwise, used for number of threads in threadpool loading,'
                  ' e.g. in vqa')
+        parlai.add_argument(
+            '--hide-labels', default=False, type='bool',
+            help='default (False) moves labels in valid and test sets to the '
+                 'eval_labels field. If True, they are hidden completely.')
         batch = self.add_argument_group('Batching Arguments')
         batch.add_argument(
             '-bs', '--batchsize', default=1, type=int,
@@ -266,7 +270,6 @@ class ParlaiParser(argparse.ArgumentParser):
         agent = get_agent_module(model)
         try:
             if hasattr(agent, 'add_cmdline_args'):
-                print('adding commandline args')
                 agent.add_cmdline_args(self)
         except argparse.ArgumentError:
             # already added
@@ -323,8 +326,8 @@ class ParlaiParser(argparse.ArgumentParser):
         if model is not None:
             self.add_model_subargs(model)
 
+        # reset parser-level defaults over any model-level defaults
         try:
-            # reset parser-level defaults over any model-level defaults
             self.set_defaults(**self._defaults)
         except AttributeError:
             raise RuntimeError('Please file an issue on github that argparse '
