+++ projects/convai2/baselines/seq2seq/eval_ppl.py
@@ -9,7 +9,7 @@ from parlai.core.dict import DictionaryAgent
 from parlai.core.params import ParlaiParser, modelzoo_path
 from parlai.agents.seq2seq.seq2seq import Seq2seqAgent
 from projects.convai2.build_dict import build_dict
-from projects.convai2.eval_ppl import setup_args, eval_ppl
+from projects.convai2.eval_ppl import eval_ppl
 import torch.nn.functional as F
 
 
@@ -17,7 +17,7 @@ class Seq2seqEntry(Seq2seqAgent):
     def __init__(self, opt, shared=None):
         super().__init__(opt, shared)
         if shared:
-            self.probs[shared['probs']]
+            self.probs = shared['probs']
         else:
             # default minimum probability mass for all tokens
             self.probs = {k: 1e-7 for k in build_dict().keys()}
@@ -74,9 +74,9 @@ class Seq2seqEntry(Seq2seqAgent):
 
 
 if __name__ == '__main__':
-    parser = setup_args()
+    parser = ParlaiParser(True, True)
     parser.set_defaults(
-        model='projects.convai2.baselines.seq2seq.seq2seq_ppl:Seq2seqEntry',
+        model='projects.convai2.baselines.seq2seq.eval_ppl:Seq2seqEntry',
         model_file='models:convai2/seq2seq/convai2_self_seq2seq_model',
         dict_file='models:convai2/seq2seq/dict_convai2_self',
         dict_lower=True,
