+++ projects/convai2/eval_f1.py
@@ -14,6 +14,7 @@ def eval_model(parser):
     parser.set_defaults(
         task='convai2:self:no_cands',
         datatype='valid',
+        hide_labels=True,
     )
     return main_eval_model(parser)
